version that passes francinette

+++ ft_strlen.c
@@ -3,21 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmillier <nmillier@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ydred <ydred@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/17 15:31:42 by nmillier          #+#    #+#             */
-/*   Updated: 2023/10/25 14:52:50 by nmillier         ###   ########.fr       */
+/*   Updated: 2023/11/03 21:22:44 by ydred            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-    int length;
+	size_t	length;
 
-    length = 0;
-    while (str[length])
-        length++;
-    return (length);
+	length = 0;
+	while (str[length])
+		length++;
+	return (length);
 }
